requirements.txt
langchain
openai
pinecone-client
langchain-pinecone
langchain-openai
python-dotenv
pypdf
langchain-huggingface
langchain_community
langchain_mistralai
faiss-cpu
numpy
mistralai
llama-index